Module google_cognitive_apis::api::grpc::google::cloud::dialogflow::cx::v3beta1

Modules

entity_type

Nested message and enum types in EntityType.

entity_types_client

Generated client implementations.

event_handler

Nested message and enum types in EventHandler.

flows_client

Generated client implementations.

form

Nested message and enum types in Form.

fulfillment

Nested message and enum types in Fulfillment.

intent

Nested message and enum types in Intent.

intents_client

Generated client implementations.

match

Nested message and enum types in Match.

match_intent_response

Nested message and enum types in MatchIntentResponse.

nlu_settings

Nested message and enum types in NluSettings.

pages_client

Generated client implementations.

query_input

Nested message and enum types in QueryInput.

query_result

Nested message and enum types in QueryResult.

response_message

Nested message and enum types in ResponseMessage.

session_entity_type

Nested message and enum types in SessionEntityType.

session_entity_types_client

Generated client implementations.

sessions_client

Generated client implementations.

streaming_detect_intent_response

Nested message and enum types in StreamingDetectIntentResponse.

streaming_recognition_result

Nested message and enum types in StreamingRecognitionResult.

transition_route

Nested message and enum types in TransitionRoute.

validation_message

Nested message and enum types in ValidationMessage.

Structs

AudioInput

Represents the natural speech audio to be processed.

CreateEntityTypeRequest

The request message for [EntityTypes.CreateEntityType][google.cloud.dialogflow.cx.v3beta1.EntityTypes.CreateEntityType].

CreateFlowRequest

The request message for [Flows.CreateFlow][google.cloud.dialogflow.cx.v3beta1.Flows.CreateFlow].

CreateIntentRequest

The request message for [Intents.CreateIntent][google.cloud.dialogflow.cx.v3beta1.Intents.CreateIntent].

CreatePageRequest

The request message for [Pages.CreatePage][google.cloud.dialogflow.cx.v3beta1.Pages.CreatePage].

CreateSessionEntityTypeRequest

The request message for [SessionEntityTypes.CreateSessionEntityType][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.CreateSessionEntityType].

DeleteEntityTypeRequest

The request message for [EntityTypes.DeleteEntityType][google.cloud.dialogflow.cx.v3beta1.EntityTypes.DeleteEntityType].

DeleteFlowRequest

The request message for [Flows.DeleteFlow][google.cloud.dialogflow.cx.v3beta1.Flows.DeleteFlow].

DeleteIntentRequest

The request message for [Intents.DeleteIntent][google.cloud.dialogflow.cx.v3beta1.Intents.DeleteIntent].

DeletePageRequest

The request message for [Pages.DeletePage][google.cloud.dialogflow.cx.v3beta1.Pages.DeletePage].

DeleteSessionEntityTypeRequest

The request message for [SessionEntityTypes.DeleteSessionEntityType][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.DeleteSessionEntityType].

DetectIntentRequest

The request to detect user’s intent.

DetectIntentResponse

The message returned from the DetectIntent method.

DtmfInput

Represents the input for dtmf event.

EntityType

Entities are extracted from user input and represent parameters that are meaningful to your application. For example, a date range, a proper name such as a geographic location or landmark, and so on. Entities represent actionable data for your application.

EventHandler

An event handler specifies an [event][google.cloud.dialogflow.cx.v3beta1.EventHandler.event] that can be handled during a session. When the specified event happens, the following actions are taken in order:

EventInput

Represents the event to trigger.

Flow

Flows represents the conversation flows when you build your chatbot agent.

FlowValidationResult

The response message for [Flows.GetFlowValidationResult][google.cloud.dialogflow.cx.v3beta1.Flows.GetFlowValidationResult].

Form

A form is a data model that groups related parameters that can be collected from the user. The process in which the agent prompts the user and collects parameter values from the user is called form filling. A form can be added to a [page][google.cloud.dialogflow.cx.v3beta1.Page]. When form filling is done, the filled parameters will be written to the [session][google.cloud.dialogflow.cx.v3beta1.SessionInfo.parameters].

FulfillIntentRequest

Request of [FulfillIntent][]

FulfillIntentResponse

Response of [FulfillIntent][]

Fulfillment

A fulfillment can do one or more of the following actions at the same time:

GetEntityTypeRequest

The request message for [EntityTypes.GetEntityType][google.cloud.dialogflow.cx.v3beta1.EntityTypes.GetEntityType].

GetFlowRequest

The response message for [Flows.GetFlow][google.cloud.dialogflow.cx.v3beta1.Flows.GetFlow].

GetFlowValidationResultRequest

The request message for [Flows.GetFlowValidationResult][google.cloud.dialogflow.cx.v3beta1.Flows.GetFlowValidationResult].

GetIntentRequest

The request message for [Intents.GetIntent][google.cloud.dialogflow.cx.v3beta1.Intents.GetIntent].

GetPageRequest

The request message for [Pages.GetPage][google.cloud.dialogflow.cx.v3beta1.Pages.GetPage].

GetSessionEntityTypeRequest

The request message for [SessionEntityTypes.GetSessionEntityType][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.GetSessionEntityType].

InputAudioConfig

Instructs the speech recognizer on how to process the audio content.

Intent

An intent represents a user’s intent to interact with a conversational agent.

IntentInput

Represents the intent to trigger programmatically rather than as a result of natural language processing.

ListEntityTypesRequest

The request message for [EntityTypes.ListEntityTypes][google.cloud.dialogflow.cx.v3beta1.EntityTypes.ListEntityTypes].

ListEntityTypesResponse

The response message for [EntityTypes.ListEntityTypes][google.cloud.dialogflow.cx.v3beta1.EntityTypes.ListEntityTypes].

ListFlowsRequest

The request message for [Flows.ListFlows][google.cloud.dialogflow.cx.v3beta1.Flows.ListFlows].

ListFlowsResponse

The response message for [Flows.ListFlows][google.cloud.dialogflow.cx.v3beta1.Flows.ListFlows].

ListIntentsRequest

The request message for [Intents.ListIntents][google.cloud.dialogflow.cx.v3beta1.Intents.ListIntents].

ListIntentsResponse

The response message for [Intents.ListIntents][google.cloud.dialogflow.cx.v3beta1.Intents.ListIntents].

ListPagesRequest

The request message for [Pages.ListPages][google.cloud.dialogflow.cx.v3beta1.Pages.ListPages].

ListPagesResponse

The response message for [Pages.ListPages][google.cloud.dialogflow.cx.v3beta1.Pages.ListPages].

ListSessionEntityTypesRequest

The request message for [SessionEntityTypes.ListSessionEntityTypes][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.ListSessionEntityTypes].

ListSessionEntityTypesResponse

The response message for [SessionEntityTypes.ListSessionEntityTypes][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.ListSessionEntityTypes].

Match

Represents one match result of [MatchIntent][].

MatchIntentRequest

Request of [MatchIntent][].

MatchIntentResponse

Response of [MatchIntent][].

NluSettings

Settings related to NLU.

OutputAudioConfig

Instructs the speech synthesizer how to generate the output audio content.

Page

A Dialogflow CX conversation (session) can be described and visualized as a state machine. The states of a CX session are represented by pages.

QueryInput

Represents the query input. It can contain one of:

QueryParameters

Represents the parameters of a conversational query.

QueryResult

Represents the result of a conversational query.

ResourceName

Resource name and display name.

ResponseMessage

Represents a response message that can be returned by a conversational agent.

SentimentAnalysisResult

The result of sentiment analysis. Sentiment analysis inspects user input and identifies the prevailing subjective opinion, especially to determine a user’s attitude as positive, negative, or neutral.

SessionEntityType

Session entity types are referred to as User entity types and are entities that are built for an individual user such as favorites, preferences, playlists, and so on.

SpeechWordInfo

Information for a word recognized by the speech recognizer.

StreamingDetectIntentRequest

The top-level message sent by the client to the [Sessions.StreamingDetectIntent][google.cloud.dialogflow.cx.v3beta1.Sessions.StreamingDetectIntent] method.

StreamingDetectIntentResponse

The top-level message returned from the StreamingDetectIntent method.

StreamingRecognitionResult

Contains a speech recognition result corresponding to a portion of the audio that is currently being processed or an indication that this is the end of the single requested utterance.

SynthesizeSpeechConfig

Configuration of how speech should be synthesized.

TextInput

Represents the natural language text to be processed.

TrainFlowRequest

The request message for [Flows.TrainFlow][google.cloud.dialogflow.cx.v3beta1.Flows.TrainFlow].

TransitionRoute

A transition route specifies a [intent][google.cloud.dialogflow.cx.v3beta1.Intent] that can be matched and/or a data condition that can be evaluated during a session. When a specified transition is matched, the following actions are taken in order:

UpdateEntityTypeRequest

The request message for [EntityTypes.UpdateEntityType][google.cloud.dialogflow.cx.v3beta1.EntityTypes.UpdateEntityType].

UpdateFlowRequest

The request message for [Flows.UpdateFlow][google.cloud.dialogflow.cx.v3beta1.Flows.UpdateFlow].

UpdateIntentRequest

The request message for [Intents.UpdateIntent][google.cloud.dialogflow.cx.v3beta1.Intents.UpdateIntent].

UpdatePageRequest

The request message for [Pages.UpdatePage][google.cloud.dialogflow.cx.v3beta1.Pages.UpdatePage].

UpdateSessionEntityTypeRequest

The request message for [SessionEntityTypes.UpdateSessionEntityType][google.cloud.dialogflow.cx.v3beta1.SessionEntityTypes.UpdateSessionEntityType].

ValidateFlowRequest

The request message for [Flows.ValidateFlow][google.cloud.dialogflow.cx.v3beta1.Flows.ValidateFlow].

ValidationMessage

Agent/flow validation message.

VoiceSelectionParams

Description of which voice to use for speech synthesis.

Enums

AudioEncoding

Audio encoding of the audio content sent in the conversational query request. Refer to the Cloud Speech API documentation for more details.

IntentView

Represents the options for views of an intent. An intent can be a sizable object. Therefore, we provide a resource view that does not return training phrases in the response.

OutputAudioEncoding

Audio encoding of the output audio format in Text-To-Speech.

SpeechModelVariant

Variant of the specified [Speech model][google.cloud.dialogflow.cx.v3beta1.InputAudioConfig.model] to use.

SsmlVoiceGender

Gender of the voice as described in SSML voice element.