Module google_cognitive_apis::api::grpc::google::cloud::dialogflow::v2

Modules

batch_update_entity_types_request

Nested message and enum types in BatchUpdateEntityTypesRequest.

batch_update_intents_request

Nested message and enum types in BatchUpdateIntentsRequest.

contexts_client

Generated client implementations.

entity_type

Nested message and enum types in EntityType.

entity_types_client

Generated client implementations.

intent

Nested message and enum types in Intent.

intents_client

Generated client implementations.

query_input

Nested message and enum types in QueryInput.

session_entity_type

Nested message and enum types in SessionEntityType.

session_entity_types_client

Generated client implementations.

sessions_client

Generated client implementations.

streaming_recognition_result

Nested message and enum types in StreamingRecognitionResult.

Structs

BatchCreateEntitiesRequest

The request message for [EntityTypes.BatchCreateEntities][google.cloud.dialogflow.v2.EntityTypes.BatchCreateEntities].

BatchDeleteEntitiesRequest

The request message for [EntityTypes.BatchDeleteEntities][google.cloud.dialogflow.v2.EntityTypes.BatchDeleteEntities].

BatchDeleteEntityTypesRequest

The request message for [EntityTypes.BatchDeleteEntityTypes][google.cloud.dialogflow.v2.EntityTypes.BatchDeleteEntityTypes].

BatchDeleteIntentsRequest

The request message for [Intents.BatchDeleteIntents][google.cloud.dialogflow.v2.Intents.BatchDeleteIntents].

BatchUpdateEntitiesRequest

The request message for [EntityTypes.BatchUpdateEntities][google.cloud.dialogflow.v2.EntityTypes.BatchUpdateEntities].

BatchUpdateEntityTypesRequest

The request message for [EntityTypes.BatchUpdateEntityTypes][google.cloud.dialogflow.v2.EntityTypes.BatchUpdateEntityTypes].

BatchUpdateEntityTypesResponse

The response message for [EntityTypes.BatchUpdateEntityTypes][google.cloud.dialogflow.v2.EntityTypes.BatchUpdateEntityTypes].

BatchUpdateIntentsRequest

BatchUpdateIntentsResponse

The response message for [Intents.BatchUpdateIntents][google.cloud.dialogflow.v2.Intents.BatchUpdateIntents].

Context

Dialogflow contexts are similar to natural language context. If a person says to you “they are orange”, you need context in order to understand what “they” is referring to. Similarly, for Dialogflow to handle an end-user expression like that, it needs to be provided with context in order to correctly match an intent.

CreateContextRequest

The request message for [Contexts.CreateContext][google.cloud.dialogflow.v2.Contexts.CreateContext].

CreateEntityTypeRequest

The request message for [EntityTypes.CreateEntityType][google.cloud.dialogflow.v2.EntityTypes.CreateEntityType].

CreateIntentRequest

The request message for [Intents.CreateIntent][google.cloud.dialogflow.v2.Intents.CreateIntent].

CreateSessionEntityTypeRequest

The request message for [SessionEntityTypes.CreateSessionEntityType][google.cloud.dialogflow.v2.SessionEntityTypes.CreateSessionEntityType].

DeleteAllContextsRequest

The request message for [Contexts.DeleteAllContexts][google.cloud.dialogflow.v2.Contexts.DeleteAllContexts].

DeleteContextRequest

The request message for [Contexts.DeleteContext][google.cloud.dialogflow.v2.Contexts.DeleteContext].

DeleteEntityTypeRequest

The request message for [EntityTypes.DeleteEntityType][google.cloud.dialogflow.v2.EntityTypes.DeleteEntityType].

DeleteIntentRequest

The request message for [Intents.DeleteIntent][google.cloud.dialogflow.v2.Intents.DeleteIntent].

DeleteSessionEntityTypeRequest

The request message for [SessionEntityTypes.DeleteSessionEntityType][google.cloud.dialogflow.v2.SessionEntityTypes.DeleteSessionEntityType].

DetectIntentRequest

The request to detect user’s intent.

DetectIntentResponse

The message returned from the DetectIntent method.

EntityType

Each intent parameter has a type, called the entity type, which dictates exactly how data from an end-user expression is extracted.

EntityTypeBatch

This message is a wrapper around a collection of entity types.

EventInput

Events allow for matching intents by event name instead of the natural language input. For instance, input <event: { name: "welcome_event", parameters: { name: "Sam" } }> can trigger a personalized welcome response. The parameter name may be used by the agent in the response: "Hello #welcome_event.name! What can I do for you today?".

GetContextRequest

The request message for [Contexts.GetContext][google.cloud.dialogflow.v2.Contexts.GetContext].

GetEntityTypeRequest

The request message for [EntityTypes.GetEntityType][google.cloud.dialogflow.v2.EntityTypes.GetEntityType].

GetIntentRequest

The request message for [Intents.GetIntent][google.cloud.dialogflow.v2.Intents.GetIntent].

GetSessionEntityTypeRequest

The request message for [SessionEntityTypes.GetSessionEntityType][google.cloud.dialogflow.v2.SessionEntityTypes.GetSessionEntityType].

InputAudioConfig

Instructs the speech recognizer how to process the audio content.

Intent

An intent categorizes an end-user’s intention for one conversation turn. For each agent, you define many intents, where your combined intents can handle a complete conversation. When an end-user writes or says something, referred to as an end-user expression or end-user input, Dialogflow matches the end-user input to the best intent in your agent. Matching an intent is also known as intent classification.

IntentBatch

This message is a wrapper around a collection of intents.

ListContextsRequest

The request message for [Contexts.ListContexts][google.cloud.dialogflow.v2.Contexts.ListContexts].

ListContextsResponse

The response message for [Contexts.ListContexts][google.cloud.dialogflow.v2.Contexts.ListContexts].

ListEntityTypesRequest

The request message for [EntityTypes.ListEntityTypes][google.cloud.dialogflow.v2.EntityTypes.ListEntityTypes].

ListEntityTypesResponse

The response message for [EntityTypes.ListEntityTypes][google.cloud.dialogflow.v2.EntityTypes.ListEntityTypes].

ListIntentsRequest

The request message for [Intents.ListIntents][google.cloud.dialogflow.v2.Intents.ListIntents].

ListIntentsResponse

The response message for [Intents.ListIntents][google.cloud.dialogflow.v2.Intents.ListIntents].

ListSessionEntityTypesRequest

The request message for [SessionEntityTypes.ListSessionEntityTypes][google.cloud.dialogflow.v2.SessionEntityTypes.ListSessionEntityTypes].

ListSessionEntityTypesResponse

The response message for [SessionEntityTypes.ListSessionEntityTypes][google.cloud.dialogflow.v2.SessionEntityTypes.ListSessionEntityTypes].

OutputAudioConfig

Instructs the speech synthesizer on how to generate the output audio content. If this audio config is supplied in a request, it overrides all existing text-to-speech settings applied to the agent.

QueryInput

Represents the query input. It can contain either:

QueryParameters

Represents the parameters of the conversational query.

QueryResult

Represents the result of conversational query or event processing.

Sentiment

The sentiment, such as positive/negative feeling or association, for a unit of analysis, such as the query text.

SentimentAnalysisRequestConfig

Configures the types of sentiment analysis to perform.

SentimentAnalysisResult

The result of sentiment analysis. Sentiment analysis inspects user input and identifies the prevailing subjective opinion, especially to determine a user’s attitude as positive, negative, or neutral. For [Participants.DetectIntent][], it needs to be configured in [DetectIntentRequest.query_params][google.cloud.dialogflow.v2.DetectIntentRequest.query_params]. For [Participants.StreamingDetectIntent][], it needs to be configured in [StreamingDetectIntentRequest.query_params][google.cloud.dialogflow.v2.StreamingDetectIntentRequest.query_params]. And for [Participants.AnalyzeContent][google.cloud.dialogflow.v2.Participants.AnalyzeContent] and [Participants.StreamingAnalyzeContent][google.cloud.dialogflow.v2.Participants.StreamingAnalyzeContent], it needs to be configured in [ConversationProfile.human_agent_assistant_config][google.cloud.dialogflow.v2.ConversationProfile.human_agent_assistant_config]

SessionEntityType

A session represents a conversation between a Dialogflow agent and an end-user. You can create special entities, called session entities, during a session. Session entities can extend or replace custom entity types and only exist during the session that they were created for. All session data, including session entities, is stored by Dialogflow for 20 minutes.

SpeechContext

Hints for the speech recognizer to help with recognition in a specific conversation state.

SpeechToTextConfig

Configures speech transcription for [ConversationProfile][google.cloud.dialogflow.v2.ConversationProfile].

SpeechWordInfo

Information for a word recognized by the speech recognizer.

StreamingDetectIntentRequest

The top-level message sent by the client to the [Sessions.StreamingDetectIntent][google.cloud.dialogflow.v2.Sessions.StreamingDetectIntent] method.

StreamingDetectIntentResponse

The top-level message returned from the StreamingDetectIntent method.

StreamingRecognitionResult

Contains a speech recognition result corresponding to a portion of the audio that is currently being processed or an indication that this is the end of the single requested utterance.

SynthesizeSpeechConfig

Configuration of how speech should be synthesized.

TextInput

Represents the natural language text to be processed.

UpdateContextRequest

The request message for [Contexts.UpdateContext][google.cloud.dialogflow.v2.Contexts.UpdateContext].

UpdateEntityTypeRequest

The request message for [EntityTypes.UpdateEntityType][google.cloud.dialogflow.v2.EntityTypes.UpdateEntityType].

UpdateIntentRequest

The request message for [Intents.UpdateIntent][google.cloud.dialogflow.v2.Intents.UpdateIntent].

UpdateSessionEntityTypeRequest

The request message for [SessionEntityTypes.UpdateSessionEntityType][google.cloud.dialogflow.v2.SessionEntityTypes.UpdateSessionEntityType].

VoiceSelectionParams

Description of which voice to use for speech synthesis.

Enums

AudioEncoding

Audio encoding of the audio content sent in the conversational query request. Refer to the Cloud Speech API documentation for more details.

IntentView

Represents the options for views of an intent. An intent can be a sizable object. Therefore, we provide a resource view that does not return training phrases in the response by default.

OutputAudioEncoding

Audio encoding of the output audio format in Text-To-Speech.

SpeechModelVariant

Variant of the specified [Speech model][google.cloud.dialogflow.v2.InputAudioConfig.model] to use.

SsmlVoiceGender

Gender of the voice as described in SSML voice element.