Struct google_cognitive_apis::api::grpc::google::cloud::dialogflow::v2beta1::QueryParameters

pub struct QueryParameters {
    pub time_zone: String,
    pub geo_location: Option<LatLng>,
    pub contexts: Vec<Context>,
    pub reset_contexts: bool,
    pub session_entity_types: Vec<SessionEntityType>,
    pub payload: Option<Struct>,
    pub knowledge_base_names: Vec<String>,
    pub sentiment_analysis_request_config: Option<SentimentAnalysisRequestConfig>,
    pub sub_agents: Vec<SubAgent>,
    pub webhook_headers: HashMap<String, String>,
}

Represents the parameters of the conversational query.

Fields
time_zone: String

The time zone of this conversational query from the time zone database, e.g., America/New_York, Europe/Paris. If not provided, the time zone specified in agent settings is used.

geo_location: Option<LatLng>

The geo location of this conversational query.

contexts: Vec<Context>

The collection of contexts to be activated before this query is executed.

reset_contexts: bool

Specifies whether to delete all contexts in the current session before the new ones are activated.

session_entity_types: Vec<SessionEntityType>

Additional session entity types to replace or extend developer entity types with. The entity synonyms apply to all languages and persist for the session of this query.

payload: Option<Struct>

This field can be used to pass custom data to your webhook. Arbitrary JSON objects are supported. If supplied, the value is used to populate the WebhookRequest.original_detect_intent_request.payload field sent to your webhook.

knowledge_base_names: Vec<String>

KnowledgeBases to get alternative results from. If not set, the KnowledgeBases enabled in the agent (through UI) will be used. Format: projects/<Project ID>/knowledgeBases/<Knowledge Base ID>.

sentiment_analysis_request_config: Option<SentimentAnalysisRequestConfig>

Configures the type of sentiment analysis to perform. If not provided, sentiment analysis is not performed. Note: Sentiment Analysis is only currently available for Essentials Edition agents.

sub_agents: Vec<SubAgent>

For mega agent query, directly specify which sub agents to query. If any specified sub agent is not linked to the mega agent, an error will be returned. If empty, Dialogflow will decide which sub agents to query. If specified for a non-mega-agent query, will be silently ignored.

webhook_headers: HashMap<String, String>

This field can be used to pass HTTP headers for a webhook call. These headers will be sent to webhook along with the headers that have been configured through Dialogflow web console. The headers defined within this field will overwrite the headers configured through Dialogflow console if there is a conflict. Header names are case-insensitive. Google’s specified headers are not allowed. Including: “Host”, “Content-Length”, “Connection”, “From”, “User-Agent”, “Accept-Encoding”, “If-Modified-Since”, “If-None-Match”, “X-Forwarded-For”, etc.