Struct google_cognitive_apis::dialogflow::sessions_client_streaming::SessionsClient

pub struct SessionsClient { /* private fields */ }

Google Dialogflow sessions client. Used for streaming detect intent API. Is NOT cloneable!

Implementations

impl SessionsClient

pub async fn create(
google_credentials: impl AsRef<str>,
streaming_detect_intent_req: StreamingDetectIntentRequest,
buffer_size: Option<usize>
) -> Result<Self>

Creates new sessions client using GCP project JSON credentials This client should be used for asynchronous invocation (streaming_detect_intent) See https://cloud.google.com/dialogflow/es/docs/how/detect-intent-stream

pub fn get_audio_sink(&mut self) -> Option<Sender<StreamingDetectIntentRequest>>

Returns sender than can be used to stream in audio bytes. This method will take the sender out of the option leaving None in its place. No additional sender can be retrieved from session client after this call. When sender is dropped respective stream will be closed.

pub fn take_audio_sink(
&mut self
) -> Option<Sender<StreamingDetectIntentRequest>>

Returns sender than can be used to stream in audio bytes. This method will take the sender out of the option leaving None in its place. No additional sender can be retrieved from session client after this call. When sender is dropped respective stream will be closed.

pub fn drop_audio_sink(&mut self)

Drops audio sender so that respective stream can be closed.

pub fn get_streaming_result_receiver(
&mut self,
buffer_size: Option<usize>
) -> Receiver<StreamingDetectIntentResponse>

Returns receiver that can be used to receive streaming detect intents results

pub fn streaming_request_from_bytes(
session: String,
audio_bytes: Vec<u8>
) -> StreamingDetectIntentRequest

Convenience function so that client does not have to create full StreamingDetectIntentRequest and can just pass audio bytes vector instead.

pub fn is_eoc(response: &DetectIntentResponse) -> bool

Convenience function to check if DetectIntentResponse represents end of conversation. If so, returns true, otherwise false.

pub fn get_session_string(project_id: &str, session_id: &str) -> String

Convenience function to return properly formatted session string for detect intent call.

pub async fn streaming_detect_intent_async_stream(
&mut self
) -> impl Stream<Item = Result<StreamingDetectIntentResponse>> + '_

IMPORTANT: currently streaming_detect_intent does not work properly because half-close operation is not implemented. Details here(go example): https://cloud.google.com/dialogflow/es/docs/how/detect-intent-stream#detect-intent-stream-go

pub async fn streaming_detect_intent(&mut self) -> Result<()>

Initiates bidirectional streaming. This call should be spawned into separate tokio task. Results can be then retrieved via channel receiver returned by method get_streaming_result_receiver. IMPORTANT: currently streaming_detect_intent does not work properly because half-close operation is not implemented. Details here(go example): https://cloud.google.com/dialogflow/es/docs/how/detect-intent-stream#detect-intent-stream-go